+++ src/app/layout.js
@@ -14,7 +14,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
 	return (
 		<html lang="en">
-			<body className={inter.className} style={{ minHeight: '100vh' }}>
+			<body className={inter.className} style={{ minHeight: '100vh', position: 'relative' }}>
 				{children}
 				<Footer />
 			</body>