+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
-## v1.0.0 (2023-xx-xx)
+## v1.0.1 (2023-06-04)
+### Feature
+* History page to show last 10 spinner history
+### Fix
+* Mobile responsive issue
+
+## v1.0.0 (2023-05-31)
 ### Release
 * Initial release